Implement sales flow, audit logging, and cache support for queries (#112)

+++ src/main/java/com/igniscore/api/dto/ProductStoreDTO.java
@@ -6,51 +6,95 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
+/**
+ * Data Transfer Object (DTO) used for product creation requests.
+ *
+ * <p>This DTO encapsulates the data required to create a new product
+ * through the API layer.
+ *
+ * <p>Main responsibilities:
+ * <ul>
+ *     <li>Transport product creation data between client and application</li>
+ *     <li>Apply declarative validation rules</li>
+ *     <li>Provide a persistence-independent request structure</li>
+ * </ul>
+ *
+ * <p>Validation rules are enforced using Jakarta Bean Validation annotations,
+ * ensuring that invalid payloads are rejected before reaching the business layer.
+ *
+ * <p>Design notes:
+ * <ul>
+ *     <li>Uses {@link LocalDate} for date-only values</li>
+ *     <li>Uses {@link BigDecimal} for monetary precision</li>
+ *     <li>Intended exclusively for create operations</li>
+ * </ul>
+ */
 public class ProductStoreDTO {
-
     /**
-     * Product name.
+     * Commercial or display name of the product.
      *
-     * <p>Must not be null or blank.
+     * <p>Validation rules:
+     * <ul>
+     *     <li>Must not be null</li>
+     *     <li>Must not be blank</li>
+     * </ul>
      */
     @NotBlank(message = "Product name is required")
     private String name;
-
     /**
-     * Product type.
+     * Product classification or category.
      *
-     * <p>Must not be null.
+     * <p>Validation rules:
+     * <ul>
+     *     <li>Must not be null</li>
+     * </ul>
      */
     @NotNull(message = "Product type is required")
     private ProductType type;
-
     /**
-     * Product expiration date.
+     * Product expiration or validity date.
      *
-     * <p>Must be a future date.
+     * <p>Validation rules:
+     * <ul>
+     *     <li>Must not be null</li>
+     *     <li>Must represent a future date</li>
+     * </ul>
+     *
+     * <p>Uses {@link LocalDate} because time-zone information
+     * is not required for this domain attribute.
      */
     @NotNull(message = "Validity date is required")
     @Future(message = "Validity date must be in the future")
     private LocalDate validity;
-
     /**
-     * Product batch/lot number.
+     * Product batch or lot identifier used for traceability.
      *
-     * <p>Must not be null or blank.
+     * <p>Validation rules:
+     * <ul>
+     *     <li>Must not be null</li>
+     *     <li>Must not be blank</li>
+     * </ul>
      */
     @NotBlank(message = "Lot is required")
     private String lot;
-
     /**
-     * Product sale price.
+     * Monetary value associated with the product.
      *
-     * <p>Must be greater than zero.
+     * <p>Validation rules:
+     * <ul>
+     *     <li>Must not be null</li>
+     *     <li>Must be greater than zero</li>
+     * </ul>
+     *
+     * <p>Uses {@link BigDecimal} to preserve decimal precision
+     * required for financial calculations.
      */
     @NotNull(message = "Price is required")
     @Positive(message = "Price must be greater than zero")
-    private Float price;
+    private BigDecimal price;
 
     public String getName() {
         return name;
@@ -68,7 +112,7 @@ public class ProductStoreDTO {
         return lot;
     }
 
-    public Float getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
@@ -88,7 +132,7 @@ public class ProductStoreDTO {
         this.lot = lot;
     }
 
-    public void setPrice(Float price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
